requirements.txt
pytest==8.2.2
pyyaml==6.0.1
openai==2.8.1
jinja2==3.1.4
python-dotenv==1.0.0
langchain==0.3.27
langchain-openai==0.3.35